+++ PayPalNativePayments/src/main/java/com/paypal/android/paypalnativepayments/PayPalNativeCheckoutResult.kt
@@ -1,10 +1,15 @@
 package com.paypal.android.paypalnativepayments
 
-import com.paypal.checkout.approve.Approval
-
 /**
  * A result passed to a [PayPalNativeCheckoutListener] when the PayPal flow completes successfully.
  */
 data class PayPalNativeCheckoutResult(
-    val approval: Approval,
+    /**
+     *  The order ID associated with the transaction
+     */
+    val orderID: String?,
+    /**
+     * The Payer ID (or user ID) associated with the transaction
+     */
+    val payerID: String?
 )