+++ PayPalNativePayments/src/test/java/com/paypal/android/paypalnativepayments/PayPalNativeCheckoutClientTest.kt
@@ -10,7 +10,6 @@ import com.paypal.checkout.approve.Approval
 import com.paypal.checkout.approve.OnApprove
 import com.paypal.checkout.cancel.OnCancel
 import com.paypal.checkout.config.CheckoutConfig
-import com.paypal.checkout.createorder.CreateOrder
 import com.paypal.checkout.error.ErrorInfo
 import com.paypal.checkout.error.OnError
 import com.paypal.checkout.shipping.OnShippingChange
@@ -209,17 +208,17 @@ class PayPalNativeCheckoutClientTest {
 
     @Test
     fun `when startCheckout is invoked, PayPalCheckout startCheckout is called`() = runTest {
-        val createOrder = mockk<CreateOrder>(relaxed = true)
+        val request = PayPalNativeCheckoutRequest("mock_order_id")
         every { PayPalCheckout.startCheckout(any()) } just runs
 
         sut = getPayPalCheckoutClient(testScheduler = testScheduler)
         resetField(PayPalCheckout::class.java, "isConfigSet", true)
 
-        sut.startCheckout(createOrder)
+        sut.startCheckout(request)
         advanceUntilIdle()
 
         verify {
-            PayPalCheckout.startCheckout(createOrder)
+            PayPalCheckout.startCheckout(any())
         }
     }
 
@@ -241,7 +240,11 @@ class PayPalNativeCheckoutClientTest {
 
     @Test
     fun `when OnApprove is invoked, onPayPalSuccess is called`() {
-        val approval = mockk<Approval>()
+        val mockOrderID = "mock_order_id"
+        val mockPayerID = "mock_payer_id"
+        val approval = mockk<Approval>(relaxed = true)
+        every { approval.data.payerId } returns mockPayerID
+        every { approval.data.orderId } returns mockOrderID
         val onApproveSlot = slot<OnApprove>()
         val paypalCheckoutResultSlot = slot<PayPalNativeCheckoutResult>()
 
@@ -262,7 +265,8 @@ class PayPalNativeCheckoutClientTest {
         every {
             payPalClientListener.onPayPalCheckoutSuccess(capture(paypalCheckoutResultSlot))
         } answers {
-            assert(paypalCheckoutResultSlot.captured.approval == approval)
+            assert(paypalCheckoutResultSlot.captured.orderID == mockOrderID)
+            assert(paypalCheckoutResultSlot.captured.payerID == mockOrderID)
         }
 
         verify { payPalClientListener.onPayPalCheckoutSuccess(any()) }